@@ -1,14 +1,16 @@
 set(TARGET DobieSDL)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 find_package(SDL2 REQUIRED)
 
 
 set(SOURCES
+	bios.cpp
     main.cpp)
 
-set(HEADERS "")
+set(HEADERS
+	bios.hpp)
 
 add_executable(${TARGET} ${SOURCES} ${HEADERS})
 target_compile_options(${TARGET} PRIVATE ${DOBIE_FLAGS})
